Create reports of a table

Create reports from data base tables, R table objects (data.table, data.frame, tibble)

dbreport(
  con = NULL,
  schema = NULL,
  tbl = NULL,
  col = NULL,
  entry = NULL,
  title = "My report title",
  text = NULL,
  plot_distinct = TRUE,
  plot_type = c("lollipop", "bar_horiz", "treemap"),
  plot_limit = 40L,
  output_dir = NULL,
  output_file = "report",
  output_format = c("html_document", "html_vignette", "pdf_document", "word_document"),
  file = TRUE,
  file_format = c("csv", "json"),
  file_type = c("single", "multiple"),
  exit = TRUE,
  verbose = FALSE
)

Arguments

con
Connection to data base, R-object (data.table, data.frame, tibble). Required.

schema
Data base schema (if applicable).

tbl
Table to be used for summary report (if applicable).

col
Table columns that should be summarised in report.

entry
Table entries which should be comsidered in summary report.

title
Title of summary report.

text
User defined text input to add some description to the report. Default NULL - no text. Can be a character vector or a file.

plot_distinct
Include plots of ordered distinct counts of each column in summary report.

plot_type
Should a lollipop or a treemap plot be produced?

plot_limit
Limit distinct count axis in the plots. Default: 100L. Should be set to a reasonable size.

output_dir
To which directory should the output be saved?

output_file
File name

output_format
Which markdown output_format should be used for rendering the summary report? E.g. html_document, html_vignette, pdf_document, ...

file
Should one or multiple files for each distinct column counts br created?

file_format
To which format should the file(s) be put out (csv, json). Multiple possible.

file_type
Output each column distinct counts set to a single or to multiple files.

exit
logical. Should database connections be closed after runing dbreport(). Default: TRUE.

verbose
TODO

Value

The function creates a markdown report and exports files

Examples

# \donttest{
# creates a .html, .pdf, .doc via Rmarkdown
dbreport(con = iris,
         plot_distinct = TRUE,
         plot_type = 'lollipop',
         plot_limit = 10, # Limit your plot to the first 15 entries
         output_dir = file.path(tempdir(), 'iris'),
         output_file = 'test',
         output_format = 'pdf_document', # you can also create a .pdf
         title = 'My iris report',
         file = TRUE, # Output single distinct counts to a file
         file_type = 'multiple',
         file_format = 'csv')

#> 
#> 
#> processing file: rmarkdown_template.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>    inline R code fragments
#> 
#>   |                                                                              |..........                                                            |  14%
#> label: data (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#> 
#>   |                                                                              |...............                                                       |  21%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................                                                  |  29%
#> label: text (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.........................                                             |  36%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................                                        |  43%
#> label: meta (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#> 
#>   |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................                              |  57%
#> label: table (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#> 
#>   |                                                                              |.............................................                         |  64%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................................                    |  71%
#> label: plot (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ fig.height: num 7
#>  $ fig.width : num 15
#>  $ dpi       : num 72
#>  $ message   : logi FALSE
#>  $ warning   : logi FALSE
#>  $ results   : chr "asis"
#> 

#> Warning: Removed 7 rows containing missing values (geom_segment).

#> Warning: Removed 7 rows containing missing values (geom_point).

#> Warning: Removed 7 rows containing missing values (geom_text).

#>   |                                                                              |.......................................................               |  79%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................................................          |  86%
#> label: plot-OLD (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ fig.height: num 7
#>  $ fig.width : num 15
#>  $ dpi       : num 72
#>  $ message   : logi FALSE
#>  $ warning   : logi FALSE
#>  $ results   : chr "asis"
#> 
#>   |                                                                              |.................................................................     |  93%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: write (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ message: logi FALSE
#>  $ warning: logi FALSE
#> 
#> 

#> output file: rmarkdown_template.knit.md

#> /usr/lib/rstudio/bin/pandoc/pandoc +RTS -K512m -RTS rmarkdown_template.utf8.md --to latex --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpZihX1f/iris/test.tex --self-contained --table-of-contents --toc-depth 3 --highlight-style tango --pdf-engine pdflatex --variable graphics --lua-filter /home/scharmueller/R/x86_64-pc-linux-gnu-library/3.6/rmarkdown/rmd/lua/pagebreak.lua --lua-filter /home/scharmueller/R/x86_64-pc-linux-gnu-library/3.6/rmarkdown/rmd/lua/latex-div.lua --variable 'geometry:margin=1in' 

#> 
#> Output created: /tmp/RtmpZihX1f/iris/test.pdf

# }